Event Time System
`Game:addPromise(promise, event_time)` - Add the given promise. `promise` must be a cutscene in the `promises.lua` file (add it to the main scripts please) and `event_time` is the amount of event time a promise takes.

`Game:addEventTime(time_added)` - Adds `time_added` event time to all promises and calls the cutscenes for them. Call this in a `cutscene.after()`, otherwise, it will crash.