Notsuki did Karaoke!
There is now an odd recording of Notsuki singing an obscure vocaloid song in the jukebox.
Despite her hating the performance, Anna somehow managed to get it into the Jukebox.
(I added this to make the world feel a bit more alive, like this sign that Notsuki lives is somewhere else now.)